Maintain first instance index/position of templates in assets file.
Grep the template paths and get the position of the first template, shifting off first template to create a list of paths to not set URLs for.

Set URLs for all paths, minus the remaining template paths (templates having one aggregate URL).

If necessary, set the asset_url at the index of the first template path (this maintains our position in the yml file).