Add fingerprints support
You can turn it on by seeting 'fingerprints_enabled: true' in your
assets.yml.

If it is enabled, when you package your assets, the fingerprint suffix
will be added to them, and assets.yaml.lock will be created. Then, when
the pages are displayed, the values from assets.yaml.lock will be
retrieved to get the suffix for the asset.